Initial register page route work
회원가입 페이지를 위한 라우트 작업입니다.

랜딩 페이지 내에서 router-view를 이용해 로그인-회원가입 뷰를
부드럽게 전환합니다.
