+++ Web/src/pages/landing/LandingRegisterView.vue
@@ -7,7 +7,7 @@ import Vue from "vue";
 import Component from "vue-class-component";
 
 @Component
-export default class RegisterPage extends Vue {
+export default class LandingRegisterView extends Vue {
 
 }
 </script>
